Revert "Fix gcc-4.7.4 and above led to device booting error"
This reverts commit c31afb2e2e22786afe38390601b9b8aa7fe4453a.

This issue has been fixed in Linux 3.4.112:
Patch:
https://git.kernel.org/cgit/linux/kernel/git/lizf/linux-3.4.y.git/commit/arch/arm/Makefile?h=linux-3.4.y&id=59463bb2d1c003c1b6618c74ad319996f732d82d

The patch turns off GCC SRA optimization, which handles promotion of signed
struct members incorrectly.

VRP (Value Range Propagation) optimization may not be the cause of device
failure to booting up, so I do the revert job. But I'm not testing with
the patch, because I no longer have this device.

I suggest if you want to compile the kernel, please use GCC 4.6.x,
because the newer version of GCC is not fully backward compatible,
it's even more so to deal with the underlying code.

Signed-off-by: Andy Deng <theandy.deng@gmail.com>